src/main/java/com/springjpa/study/jpashop/repo/MemberRepository.java
package com.springjpa.study.jpashop.repo;

import com.springjpa.study.jpashop.domain.Member;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

/**
 * 회원 객체 레포지토리.
 *
 * @see com.springjpa.study.jpashop.domain.Member
 */
@Repository
public class MemberRepository {

    @PersistenceContext
    private EntityManager em;

    public Long save(Member member) {
        em.persist(member);
        return member.getId();
    }

    public Member find(Long id) {
        return em.find(Member.class, id);
    }
}